+++ frontend/src/App.js
@@ -20,6 +20,7 @@ import VoiceAssistant from './components/VoiceAssistant';
 import ClientWelcome from './pages/client/ClientWelcome';
 import UserHome from './pages/user/UserHome';
 import RideBookingPage from './pages/user/RideBookingPage';
+import RideChoosePage from './pages/user/RideChoosePage';
 import RideTrackingPage from './pages/user/RideTrackingPage';
 import RideReceiptPage from './pages/user/RideReceiptPage';
 import FoodPage from './pages/user/FoodPage';
@@ -274,6 +275,7 @@ const AppRouter = () => {
       <Route path="/taxi-bidding" element={<ProtectedRoute allowedRoles={['user']}><TaxiBiddingPage /></ProtectedRoute>} />
       <Route path="/taxi-advanced" element={<ProtectedRoute allowedRoles={['user']}><AdvancedTaxiBookingPage /></ProtectedRoute>} />
       <Route path="/taxi" element={<ProtectedRoute allowedRoles={['user']}><TaxiHubPage /></ProtectedRoute>} />
+      <Route path="/course" element={<ProtectedRoute allowedRoles={['user']}><RideChoosePage /></ProtectedRoute>} />
       <Route path="/scheduled-rides" element={<ProtectedRoute allowedRoles={['user']}><ScheduledRidesPage /></ProtectedRoute>} />
       <Route path="/corporate" element={<ProtectedRoute allowedRoles={['user']}><CorporateAccountPage /></ProtectedRoute>} />
       <Route path="/runner" element={<ProtectedRoute allowedRoles={['user']}><RunnerPage /></ProtectedRoute>} />